requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-filter==2.4.0
Pillow==8.0.1
kombu==4.6.10
celery==4.4.7
requests
opencv-python
mysqlclient==2.0.1
mysql-connector
mysql-connector-python
django-mysql